chore: ruff + mypy 도입 — 전 소스 클린, CI 에 린트 단계 추가
- ruff (기본 + SIM/BLE/PLW/RUF): 미사용 임포트 15파일, 임포트 정렬,
  불필요한 int(round()) 등 48건 정리. 의도적 광역 except(가용성
  프로브)는 사유 주석과 함께 noqa
- mypy: 실제 문제 3건 수정 — trace 위젯의 루프 변수 섀도잉(float 가
  float|None 으로 재사용), QSettings.value 반환 타입 미검증 2곳
- pyproject 에 설정 고정, dev extras 에 추가, CI 는 테스트 전에
  ruff/mypy 를 먼저 돌린다

+++ src/tuner/analysis/reference.py
@@ -49,7 +49,7 @@ def annotate(
     Each estimate uses a long analysis frame centered on the slice midpoint
     (the offline advantage: future samples are available).
     """
-    hop = int(round(window_s * sr))
+    hop = round(window_s * sr)
     frame_size = max(MIN_ANALYSIS_FRAME, 4 * hop)
     n_windows = len(signal) // hop
     window_rms = np.array(
